// src/shared/enums/userPreferenceKeys.ts
export enum USER_PREFERENCE_KEYS {
  MOD_DIRECTORY = "MOD_DIRECTORY",
  PRESET = "PRESET",
  GRAPHICS = "GRAPHICS",
  ENB_PROFILE = "ENB_PROFILE",
  PREVIOUS_ENB_PROFILE = "PREVIOUS_ENB_PROFILE",
  RESOLUTION = "RESOLUTION",
  SHOW_HIDDEN_PROFILE = "SHOW_HIDDEN_PROFILE",

  CHECK_PREREQUISITES = "CHECK_PREREQUISITES",
}